move node_killer to spec folder
