+++ sh/env.rb
@@ -4,4 +4,4 @@ require 'etcd'
 require 'json'
 
 ## only for testing in the console
-require 'etcd/node_killer'
+require './spec/resources/node_killer'
