+++ spec/resources/node_killer.rb
